+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
-- (Enhancement) Add support for Intrusion Rule Group
-- (Enhancement) Add support for Intrusion Rules
+- (Enhancement) Add support for Intrusion Rules and Intrusion Rule Group
+- (Enhancement) Add support for Intrusion Policy Group Overrides
 
 ## 2.2.0